issue-524 Step 5: extract connection management and schema migration
Adds SqlitePersistenceAdapter in control_plane/adapters.py — a verbatim
behavior move of ControlPlane's connection management (_get_connection,
_discover_shared_db_path) and schema migration engine (init_db,
_schema_needs_rebuild, _rebuild_schema_transactional,
_copy_common_columns, _merge_orphaned_tasks_old, _log_orphan_merge_conflicts,
plus SCHEMA_SQL/SCHEMA_MIGRATIONS/CURRENT_SCHEMA_VERSION/CHILD_TABLES/
ALL_REBUILD_TABLES). ControlPlane.__init__ now composes
self._persistence = SqlitePersistenceAdapter(...); _get_connection()/
init_db() are one-line delegations. self.db_path kept as a
backward-compatible convenience mirror.

Fixed test_control_plane_uses_injected_filesystem_port_for_orphan_merge_log
(added Step 3), which called _log_orphan_merge_conflicts directly on
ControlPlane and broke once that method legitimately relocated to the
persistence adapter — confirmed as expected relocation, not new dead
code. Removed the now-orphaned `import subprocess` from agent_control.py.

138/139 tests passing repo-wide (1 pre-existing unrelated skip). New: 5
adapter-level unit tests exercising SqlitePersistenceAdapter directly.
Structural audit clean.
