+++ plugins/agent-agentic-os/references/evolution-log.md
@@ -2,6 +2,7 @@
 
 | Date | Tier | Friction / Failure | Patch | Edit Type | Outcome |
 |------|------|-------------------|-------|-----------|---------|
+| 2026-09-06 | Tier 2/3 (in progress) | issue-524 Step 5: connection management (`_get_connection`, `_discover_shared_db_path`) and the schema migration engine (`init_db`, `_schema_needs_rebuild`, `_rebuild_schema_transactional`, `_copy_common_columns`, `_merge_orphaned_tasks_old`, plus `SCHEMA_SQL`/`SCHEMA_MIGRATIONS`/`CURRENT_SCHEMA_VERSION`/`CHILD_TABLES`/`ALL_REBUILD_TABLES`) were mixed directly into `ControlPlane`, one of the 7 god-object responsibilities. A pre-existing test (`test_control_plane_uses_injected_filesystem_port_for_orphan_merge_log`, added Step 3) called `_log_orphan_merge_conflicts` directly on `ControlPlane` and broke once that method legitimately moved to the persistence adapter. | Added `SqlitePersistenceAdapter` in `control_plane/adapters.py` — verbatim behavior move of all connection/migration logic, including `_log_orphan_merge_conflicts` (composes `FilesystemPort`, preserving the Step 3 seam). `ControlPlane.__init__` now composes `self._persistence = SqlitePersistenceAdapter(...)`; `_get_connection()`/`init_db()` became one-line delegations; `self.db_path` kept as a convenience mirror for introspection. Fixed the Step 3 test to call `cp._persistence._log_orphan_merge_conflicts(...)` instead — confirmed as legitimate relocation, not new dead code. Backward-compat: `CURRENT_SCHEMA_VERSION` re-exported from `agent_control` for existing `from agent_control import CURRENT_SCHEMA_VERSION` callers. Removed the now-orphaned `import subprocess` from `agent_control.py` (moved with the discovery logic). | Fix Forward (in progress) | 138/139 tests passing repo-wide (1 pre-existing unrelated skip). New: 5 adapter-level unit tests. Structural audit clean. |
 | 2026-09-06 | Tier 2/3 (in progress) | issue-524 Step 4 (the plan's highest-risk step): gate policy was scattered across 4 independent mechanisms — the `GATE_REQUIREMENTS` registry, 3 hardcoded `_check_*_guard` methods, and `update_worktree()`'s own `pushed_to_origin` barrier — with the push barrier structurally unable to fit the transition-keyed registry (it's an operation guard, not a `(from_state, to_state)` edge), per external plan review round 1. | Added `control_plane/policy.py`: one declarative engine evaluating two rule kinds — `TRANSITION_RULES`/`TO_STATE_RULES` (lifecycle transitions, formerly `GATE_REQUIREMENTS` + the 3 guards) and `OPERATION_RULES` (controlled operations, formerly the push barrier) — both raising the same `PolicyViolation`, re-raised as `PersistenceInvariantViolation` in `agent_control.py` to preserve the exact CLI exit-code contract. Removed all 4 old mechanisms and their helper methods entirely (`_check_prior_art_guard`, `_check_done_guard`, `_check_rolled_back_guard`, `_check_gate_requirement`, `_gate_receipt_exists`, `_gate_critic_review_pass_exists`, `_gate_any_of`, `GATE_REQUIREMENTS`) — confirmed via a permanent AST-based structural test, not just a manual grep. | Fix Forward (in progress) | 133/134 tests passing repo-wide (1 pre-existing unrelated skip) — the Step 1 characterization matrix (38 tests) stayed green throughout, proving behavior preservation across the riskiest step. New: 11 policy-engine unit tests running in 0.03s with zero SQLite fixtures — the exact testability benefit the issue asked for. |
 | 2026-09-06 | Tier 2/3 (in progress) | issue-524 Step 3: `_log_orphan_merge_conflicts()` wrote directly to `references/map-debt.md` via a raw `open()`, and (discovered while wiring the seam) `test_control_plane_worktree_sharing.py`'s fixture copied only `agent_control.py` into its throwaway repo — a pattern that would have silently broken once `agent_control.py` gained a real dependency on the new `control_plane/` subpackage. | Added `FilesystemAdapter` in `control_plane/adapters.py` implementing `FilesystemPort`; `ControlPlane.__init__` now accepts an optional `fs_adapter` (defaults to `FilesystemAdapter()`, preserving existing behavior for all current callers) and `_log_orphan_merge_conflicts()` routes through `self._fs` instead of a raw `open()`. Fixed the stale test fixture to `copytree` the `control_plane/` package alongside `agent_control.py` — confirmed this is fixture staleness, not a production risk, since the real plugin installer (`plugin_installer.py`) already deploys whole directories via `copytree`. | Fix Forward (in progress) | 122/123 tests passing repo-wide (1 pre-existing unrelated skip), zero regressions. New: 46 tests in the issue-524 test files (up from 38). |
 | 2026-09-06 | Tier 2/3 (in progress) | issue-524 Step 2: no ports/interfaces existed to separate domain/policy logic from concrete infrastructure (sqlite3, subprocess, hashlib, filesystem, model-catalog JSON storage) inside `ControlPlane`. | Added `plugins/agent-agentic-os/scripts/control_plane/` package with `ports.py` declaring `PersistencePort`, `FilesystemPort`, `CryptoPort`, `ClockPort`, `ModelCatalogPort` as ABCs. Added `test_control_plane_ports.py` with a static AST-based check enforcing the dependency-direction invariant (no sqlite3/subprocess/hashlib/time/os imports in the ports module) plus ABC-instantiation and standalone-import checks. Purely additive — nothing wired in yet, no behavior change. | Fix Forward (in progress) | 41/41 tests passing (38 prior + 3 new). |
