+++ plugins/agent-agentic-os/scripts/agent_control.py
@@ -65,14 +65,13 @@ import hashlib
 import json
 import os
 import sqlite3
-import subprocess
 import sys
 import time
 from pathlib import Path
 from typing import Dict, List, Optional, Any
 
 from control_plane.ports import FilesystemPort
-from control_plane.adapters import FilesystemAdapter
+from control_plane.adapters import FilesystemAdapter, SqlitePersistenceAdapter, CURRENT_SCHEMA_VERSION
 from control_plane import policy as _policy
 
 CANONICAL_STATES = [
@@ -118,108 +117,10 @@ ALLOWED_TRANSITIONS = {
     "ESCALATED": ["INTAKE", "PLAN_REVIEW"]
 }
 
-SCHEMA_SQL = """
-PRAGMA foreign_keys = ON;
-PRAGMA journal_mode = WAL;
-PRAGMA busy_timeout = 5000;
-
-CREATE TABLE IF NOT EXISTS tasks (
-    task_id TEXT PRIMARY KEY,
-    title TEXT NOT NULL,
-    state TEXT NOT NULL CHECK (
-        state IN (
-            'INTAKE', 'INTERVIEW', 'DRAFT_PLAN', 'MULTI_AGENT_REVIEW', 'PLAN_REVIEW', 'AWAITING_APPROVAL',
-            'APPROVED', 'IN_WORKTREE', 'WORKTREE_REVIEW', 'MULTI_AGENT_CODE_REVIEW', 'VERIFY_EXIT', 'DONE',
-            'ROLLED_BACK', 'ESCALATED'
-        )
-    ),
-    task_type TEXT NOT NULL DEFAULT 'GENERAL' CHECK (task_type IN ('GENERAL', 'EVOLUTION')),
-    runtime_tool TEXT NOT NULL,
-    worktree_path TEXT,
-    worktree_branch TEXT,
-    worktree_state TEXT CHECK (
-        worktree_state IS NULL OR worktree_state IN (
-            'written_in_worktree', 'committed_in_worktree', 'pushed_to_origin',
-            'merged_into_origin_main', 'local_branch_ref_updated', 'checked_out_on_disk'
-        )
-    ),
-    spec_path TEXT,
-    model_tier TEXT CHECK (model_tier IS NULL OR model_tier IN ('low', 'medium', 'high')),
-    model_id TEXT,
-    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-);
-
-CREATE TABLE IF NOT EXISTS task_transitions (
-    transition_id INTEGER PRIMARY KEY AUTOINCREMENT,
-    task_id TEXT NOT NULL REFERENCES tasks(task_id) ON DELETE CASCADE,
-    from_state TEXT NOT NULL,
-    to_state TEXT NOT NULL,
-    actor TEXT NOT NULL,
-    reason TEXT,
-    timestamp TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-);
-
-CREATE TABLE IF NOT EXISTS locked_verifier_baselines (
-    baseline_id INTEGER PRIMARY KEY AUTOINCREMENT,
-    task_id TEXT NOT NULL REFERENCES tasks(task_id) ON DELETE CASCADE,
-    file_path TEXT NOT NULL,
-    expected_sha256 TEXT NOT NULL,
-    verified_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-);
-
-CREATE TABLE IF NOT EXISTS critic_reviews (
-    review_id INTEGER PRIMARY KEY AUTOINCREMENT,
-    task_id TEXT NOT NULL REFERENCES tasks(task_id) ON DELETE CASCADE,
-    iteration INTEGER NOT NULL CHECK(iteration >= 1),
-    model_used TEXT NOT NULL,
-    verdict TEXT NOT NULL CHECK (verdict IN ('PASS', 'REVISE', 'REJECT')),
-    critique_findings TEXT,
-    timestamp TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-);
-
-CREATE TABLE IF NOT EXISTS verification_receipts (
-    receipt_id INTEGER PRIMARY KEY AUTOINCREMENT,
-    task_id TEXT NOT NULL REFERENCES tasks(task_id) ON DELETE CASCADE,
-    gate_name TEXT NOT NULL,
-    command_executed TEXT NOT NULL,
-    exit_code INTEGER NOT NULL,
-    receipt_token TEXT NOT NULL,
-    timestamp TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-);
-
-CREATE TABLE IF NOT EXISTS asymmetric_persistence_log (
-    log_id INTEGER PRIMARY KEY AUTOINCREMENT,
-    task_id TEXT NOT NULL REFERENCES tasks(task_id) ON DELETE CASCADE,
-    destination TEXT NOT NULL,
-    status TEXT NOT NULL CHECK(status IN ('OBSERVED', 'HYPOTHESIS', 'CONFIRMED', 'RESOLVED')),
-    details TEXT,
-    timestamp TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-);
-
-CREATE TABLE IF NOT EXISTS schema_version (
-    version INTEGER NOT NULL
-);
-
-CREATE INDEX IF NOT EXISTS idx_tasks_state ON tasks(state);
-CREATE INDEX IF NOT EXISTS idx_transitions_task ON task_transitions(task_id);
-"""
-
-CURRENT_SCHEMA_VERSION = 3
-
-CHILD_TABLES = [
-    "task_transitions",
-    "locked_verifier_baselines",
-    "critic_reviews",
-    "verification_receipts",
-    "asymmetric_persistence_log",
-]
-ALL_REBUILD_TABLES = ["tasks"] + CHILD_TABLES
-
-SCHEMA_MIGRATIONS = [
-    # Migration: add task_type column if not present (for existing DBs)
-    "ALTER TABLE tasks ADD COLUMN task_type TEXT NOT NULL DEFAULT 'GENERAL' CHECK (task_type IN ('GENERAL', 'EVOLUTION'));",
-]
+# SCHEMA_SQL, CURRENT_SCHEMA_VERSION, CHILD_TABLES, ALL_REBUILD_TABLES, and SCHEMA_MIGRATIONS
+# moved to control_plane/adapters.py's SqlitePersistenceAdapter (issue-524 Step 5). This module
+# re-imports CURRENT_SCHEMA_VERSION above for backward-compatible `from agent_control import
+# CURRENT_SCHEMA_VERSION` (existing test/consumer import surface, unchanged).
 
 
 class InvalidStateTransition(Exception):
@@ -263,203 +164,28 @@ class ControlPlane:
     """Controller and state machine manager for task lifecycles."""
 
     def __init__(self, db_path: Optional[Path] = None, fs_adapter: Optional["FilesystemPort"] = None):
-        """Initializes the ControlPlane instance with database path. `fs_adapter` defaults to
-        a real-filesystem FilesystemAdapter — the optional parameter exists so tests/callers
-        can substitute a different FilesystemPort implementation without touching disk; public
-        callers relying on the default constructor signature are unaffected (backward-compatible
-        facade, per docs/plans/issue-524-spec.md Section 3)."""
-        if db_path is None:
-            self.db_path = self._discover_shared_db_path()
-        else:
-            self.db_path = Path(db_path)
-        self.db_path.parent.mkdir(parents=True, exist_ok=True)
+        """Initializes the ControlPlane instance. Connection management and schema migration
+        are delegated to a SqlitePersistenceAdapter (issue-524 Step 5) — `self.db_path` is kept
+        as a convenience property mirroring the adapter's resolved path, for any code that
+        introspects it directly. `fs_adapter` defaults to a real-filesystem FilesystemAdapter;
+        both parameters are optional so tests/callers can substitute different Port
+        implementations without touching disk — public callers relying on the default
+        constructor signature are unaffected (backward-compatible facade, per
+        docs/plans/issue-524-spec.md Section 3)."""
         self._fs = fs_adapter if fs_adapter is not None else FilesystemAdapter()
-
-    def _discover_shared_db_path(self) -> Path:
-        """Resolves context/control_plane.db anchored at the repo root shared across
-        ALL git worktrees of the same repo — via `git rev-parse --git-common-dir`,
-        which (unlike a manual walk-up for the nearest `.git`) returns the same
-        physical .git directory whether invoked from the main checkout or any
-        worktree, so they all read/write one control plane instead of each worktree
-        getting its own disconnected DB."""
-        curr = Path(__file__).resolve().parent
-        for p in [curr] + list(curr.parents):
-            if p.name == ".agents":
-                return p.parent / "context" / "control_plane.db"
-
-        try:
-            result = subprocess.run(
-                ["git", "rev-parse", "--git-common-dir"],
-                cwd=str(curr), capture_output=True, text=True, timeout=5
-            )
-            if result.returncode == 0:
-                common_dir = Path(result.stdout.strip())
-                if not common_dir.is_absolute():
-                    common_dir = (curr / common_dir).resolve()
-                return common_dir.parent / "context" / "control_plane.db"
-        except (subprocess.SubprocessError, OSError):
-            pass
-
-        for p in [curr] + list(curr.parents):
-            if (p / "context").exists() and not (p / "skills").exists():
-                return p / "context" / "control_plane.db"
-
-        return Path.cwd() / "context" / "control_plane.db"
+        self._persistence = SqlitePersistenceAdapter(
+            db_path if db_path is None else Path(db_path), self._fs
+        )
+        self.db_path = self._persistence.db_path
 
     def _get_connection(self) -> sqlite3.Connection:
-        """Returns a configured sqlite3 connection with WAL mode and foreign keys."""
-        conn = sqlite3.connect(str(self.db_path), timeout=10.0)
-        conn.row_factory = sqlite3.Row
-        conn.execute("PRAGMA foreign_keys = ON;")
-        conn.execute("PRAGMA busy_timeout = 5000;")
-        return conn
+        """Returns a configured sqlite3 connection. Delegates to the SqlitePersistenceAdapter."""
+        return self._persistence.get_connection()
 
     def init_db(self):
-        """Initializes SQLite tables and WAL mode. Self-heals FK-corrupted or legacy schemas."""
-        conn = self._get_connection()
-        try:
-            conn.execute("PRAGMA journal_mode = WAL;")
-            conn.executescript(SCHEMA_SQL)
-            # Apply idempotent migrations for existing databases
-            for migration in SCHEMA_MIGRATIONS:
-                try:
-                    conn.execute(migration)
-                    conn.commit()
-                except sqlite3.OperationalError as e:
-                    if "duplicate column" not in str(e).lower():
-                        raise
-                    # Column already exists — expected on a previously-migrated DB, ignore.
-
-            if self._schema_needs_rebuild(conn):
-                self._rebuild_schema_transactional(conn)
-            else:
-                conn.execute(
-                    "INSERT INTO schema_version (version) SELECT ? WHERE NOT EXISTS (SELECT 1 FROM schema_version)",
-                    (CURRENT_SCHEMA_VERSION,)
-                )
-                conn.commit()
-        finally:
-            conn.close()
-
-    def _schema_needs_rebuild(self, conn: sqlite3.Connection) -> bool:
-        """Detects a stale schema_version, a legacy tasks schema, or FK-corrupted/orphaned
-        migration artifacts. schema_version is the primary migration trigger; the DDL-sniff
-        and corruption checks below are defensive fallbacks for pre-schema_version databases
-        and self-healing already-corrupted ones, not the source of truth for version bumps."""
-        version_row = conn.execute("SELECT version FROM schema_version").fetchone()
-        if version_row and version_row[0] < CURRENT_SCHEMA_VERSION:
-            return True
-
-        row = conn.execute("SELECT sql FROM sqlite_master WHERE type='table' AND name='tasks'").fetchone()
-        if row and "WORKTREE_REVIEW" not in (row[0] or ""):
-            return True
-
-        orphan = conn.execute(
-            "SELECT COUNT(*) FROM sqlite_master WHERE type='table' AND name='_tasks_old'"
-        ).fetchone()[0]
-        if orphan:
-            return True
-
-        for table in CHILD_TABLES:
-            child_row = conn.execute(
-                "SELECT sql FROM sqlite_master WHERE type='table' AND name=?", (table,)
-            ).fetchone()
-            if child_row and child_row[0] and ("_tasks_old" in child_row[0] or "_migrating" in child_row[0]):
-                return True
-        return False
-
-    def _copy_common_columns(self, conn: sqlite3.Connection, source_table: str, dest_table: str):
-        """Copies rows between tables via the intersection of their columns."""
-        source_cols = [r[1] for r in conn.execute(f'PRAGMA table_info("{source_table}");').fetchall()]
-        dest_cols = [r[1] for r in conn.execute(f'PRAGMA table_info("{dest_table}");').fetchall()]
-        common_cols = [c for c in source_cols if c in dest_cols]
-        cols_str = ", ".join(f'"{c}"' for c in common_cols)
-        conn.execute(f'INSERT INTO "{dest_table}" ({cols_str}) SELECT {cols_str} FROM "{source_table}";')
-
-    def _merge_orphaned_tasks_old(self, conn: sqlite3.Connection):
-        """Merges a dangling `_tasks_old` (from a previously-interrupted migration) into the
-        fresh `tasks` table. Logs any conflicting task_id (present in both with different
-        values) to map-debt instead of silently discarding it."""
-        merge_cols = [r[1] for r in conn.execute('PRAGMA table_info("_tasks_old_merge");').fetchall()]
-        target_cols = [r[1] for r in conn.execute('PRAGMA table_info("tasks");').fetchall()]
-        common_cols = [c for c in merge_cols if c in target_cols]
-        cols_str = ", ".join(f'"{c}"' for c in common_cols)
-
-        existing_ids = {r[0] for r in conn.execute("SELECT task_id FROM tasks").fetchall()}
-        conflicts = []
-        for row in conn.execute(f'SELECT {cols_str} FROM "_tasks_old_merge"').fetchall():
-            row_dict = dict(zip(common_cols, row))
-            task_id = row_dict["task_id"]
-            if task_id in existing_ids:
-                conflicts.append(task_id)
-                continue
-            placeholders = ", ".join(["?"] * len(common_cols))
-            conn.execute(f'INSERT INTO tasks ({cols_str}) VALUES ({placeholders})', row)
-
-        if conflicts:
-            self._log_orphan_merge_conflicts(conflicts)
-
-    def _log_orphan_merge_conflicts(self, conflicting_task_ids: List[str]):
-        """Appends a map-debt entry for task_ids dropped during an orphaned-table merge
-        (present in both the fresh `tasks` table and a dangling `_tasks_old`)."""
-        repo_root = Path(__file__).resolve().parent.parent.parent.parent
-        map_debt_path = repo_root / "references" / "map-debt.md"
-        if not self._fs.exists(map_debt_path):
-            return
-        entry = (
-            f"\n| DEBT-{time.strftime('%Y%m%d')}-AUTO | Orphaned _tasks_old merge conflict "
-            f"discarded rows for task_ids: {', '.join(conflicting_task_ids)} | OPEN | Tier 1 | 1 | "
-            f"{time.strftime('%Y-%m-%d')} | Self-heal migration in init_db() found these task_ids "
-            f"in both the fresh tasks table and a dangling _tasks_old, with differing data. "
-            f"The tasks table's version was kept; _tasks_old's version was discarded. | "
-            f"Review discarded data manually if needed; _tasks_old is already dropped. |\n"
-        )
-        self._fs.append_text(map_debt_path, entry)
-
-    def _rebuild_schema_transactional(self, conn: sqlite3.Connection):
-        """Rebuilds tasks + all child tables together in one explicit transaction, so a
-        mid-sequence failure rolls back cleanly instead of leaving a corrupted intermediate
-        state. Renaming all six tables together (instead of `tasks` alone) prevents SQLite's
-        FK-auto-repoint behavior from leaving child tables pointing at a stale name."""
-        conn.execute("PRAGMA foreign_keys = OFF;")
-        conn.execute("BEGIN IMMEDIATE;")
-        try:
-            orphan_exists = conn.execute(
-                "SELECT COUNT(*) FROM sqlite_master WHERE type='table' AND name='_tasks_old'"
-            ).fetchone()[0]
-            if orphan_exists:
-                conn.execute('ALTER TABLE "_tasks_old" RENAME TO "_tasks_old_merge";')
-
-            for table in ALL_REBUILD_TABLES:
-                exists = conn.execute(
-                    "SELECT COUNT(*) FROM sqlite_master WHERE type='table' AND name=?", (table,)
-                ).fetchone()[0]
-                if exists:
-                    conn.execute(f'ALTER TABLE "{table}" RENAME TO "_{table}_migrating";')
-
-            conn.executescript(SCHEMA_SQL)
-
-            for table in ALL_REBUILD_TABLES:
-                migrating_name = f"_{table}_migrating"
-                migrating_exists = conn.execute(
-                    "SELECT COUNT(*) FROM sqlite_master WHERE type='table' AND name=?", (migrating_name,)
-                ).fetchone()[0]
-                if migrating_exists:
-                    self._copy_common_columns(conn, migrating_name, table)
-                    conn.execute(f'DROP TABLE "{migrating_name}";')
-
-            if orphan_exists:
-                self._merge_orphaned_tasks_old(conn)
-                conn.execute('DROP TABLE "_tasks_old_merge";')
-
-            conn.execute("DELETE FROM schema_version;")
-            conn.execute("INSERT INTO schema_version (version) VALUES (?);", (CURRENT_SCHEMA_VERSION,))
-            conn.execute("COMMIT;")
-        except Exception:
-            conn.execute("ROLLBACK;")
-            raise
-        finally:
-            conn.execute("PRAGMA foreign_keys = ON;")
+        """Initializes SQLite tables and WAL mode. Delegates to the SqlitePersistenceAdapter,
+        which self-heals FK-corrupted or legacy schemas."""
+        self._persistence.ensure_schema()
 
     def _resolve_tool_catalog(self, runtime_tool: str, cli_refs: Path) -> tuple:
         """Resolves tool alias and catalog file path."""
